Additional usability updates.
Disable "on-type" formatting for lsp-mode.  This seems to interfere
with Emacs indentation and causes text to become corrupted on the
corresponding line.  This shouldn't be needed with the Emacs
indentation engine anyway as Emacs will call out to the Language
Server to perform line indentation (if indentation is configured for
LSP).

Also reworked the .dir-locals.el.  This is mainly due to the fact that
Eglot somewhat arbitrarily picks a specific major mode that it looks
for directory local configurations from when multiple modes are
configured for the same LSP server.  This is documented in Emacs bug
73280.  The workaround is to use "nil" so that no matter what mode is
being used, the settings will be applied.