Workaround Corfu narrowing problem.
It appears, when used in combination with `corfu-popupinfo`, corfu may
not display the latest candidate(s), resulting in a previous candidate
being shown, even though the " *corfu*" buffer contains the correct
candidate(s).  The pop-up window is resized correctly (based on the
filtering), but the candidate(s) contained within the filtered window
are not correct, and appears to be taken from the pre-filtered set.
Furthermore, inserting this "stale" candidate causes the correct
candidate to be inserted, not the displayed one.  This doesn't always
happen, but happens frequently enough to be problematic.  Also, it
seems to only manifest when `corfu-popupinfo` is used with a delay of
0.0 (and possibly only when there is nothing for `corfu-popupinfo` to
display).  Increasing the delay to a small non-zero value (i.e., 0.01)
seems to resolve this issue.